+++ sdk/keyvault/perf-tests/keyvault-secrets/test/secretTest.ts
@@ -9,7 +9,11 @@ export abstract class SecretTest<TOptions = Record<string, unknown>> extends Per
 
   constructor() {
     super();
-    this.secretClient = new SecretClient(keyVaultUri, credential, this.configureClientOptionsCoreV1({}));
+    this.secretClient = new SecretClient(
+      keyVaultUri,
+      credential,
+      this.configureClientOptionsCoreV1({})
+    );
   }
 
   async deleteAndPurgeSecrets(...names: string[]) {