+++ sdk/schemaregistry/schema-registry/recordings/node/schemaregistryclient/recording_fails_to_get_schema_id_when_no_matching_schema_exists.js
@@ -1,6 +1,6 @@
 let nock = require('nock');
 
-module.exports.hash = "cbdfe221ea475043fb6f9368ba28a730";
+module.exports.hash = "269fc34397d7f3d879c67d9a3a4f837a";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}
 
@@ -23,19 +23,19 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
   'P3P',
   'CP="DSP CUR OTPi IND OTRi ONL FIN"',
   'x-ms-request-id',
-  '93fc3b2f-68cc-4c9e-9c83-1752eda7d701',
+  'e343d8d5-36aa-4a56-9ec6-3b67a9c72000',
   'x-ms-ests-server',
-  '2.1.11898.12 - EUS ProdSlices',
+  '2.1.12071.13 - NCUS ProdSlices',
   'Set-Cookie',
-  'fpc=ArqRweA9seFErsvWrKI3q4nmLYaiAwAAAJ7vnNgOAAAA; expires=Fri, 03-Sep-2021 20:29:21 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'fpc=ApDR4LGoVSZHiVzwfsVRCTU; expires=Sun, 24-Oct-2021 16:46:24 GMT; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
-  'esctx=AQABAAAAAAD--DLA3VO7QrddgJg7WevrbY4LwKUCLYwS17CcSOkAdi2QAM_OsuDfnP_oErRU7oKrFaBtVYJNAJdk2MGF58LG9m2O9PemEIbvobFHBlCCj6NfmqfPjYEttC9PbdnXbuz14-ycWKL9skB4RL906n0_MxA5_zMBR0TBlAb3MT_w3lf-DvH6h1OXxVwrZbzrvrsgAA; domain=.login.microsoftonline.com; path=/; secure; HttpOnly; SameSite=None',
+  'esctx=AQABAAAAAAD--DLA3VO7QrddgJg7Wevr3wtSnne4yG9qsDcm7CLlXCD8dhu8nKr6KEzpO4mL3orBlgZJLnwcwS_IHrD1sz84cjFesoRZrx9d9iYzbqeE0Ia1yWE0JMiAycTnq3cf_vX2yV1gRN0jEuulyHBijyMMey7lZ-oICpn4OT1Md4z1xkvQecUdx3OzH6BeLtmx7oYgAA; domain=.login.microsoftonline.com; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
   'x-ms-gateway-slice=estsfd; path=/; secure; samesite=none; httponly',
   'Set-Cookie',
   'stsservicecookie=estsfd; path=/; secure; samesite=none; httponly',
   'Date',
-  'Wed, 04 Aug 2021 20:29:20 GMT',
+  'Fri, 24 Sep 2021 16:46:24 GMT',
   'Content-Length',
   '980'
 ]);
@@ -58,25 +58,25 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
   'P3P',
   'CP="DSP CUR OTPi IND OTRi ONL FIN"',
   'x-ms-request-id',
-  'd49c3da3-122f-47d6-99e1-37519dcf0600',
+  '8c5f4536-b587-4192-af93-d1f7373b3a00',
   'x-ms-ests-server',
-  '2.1.11935.12 - WUS2 ProdSlices',
+  '2.1.12071.13 - SCUS ProdSlices',
   'Set-Cookie',
-  'fpc=ArqRweA9seFErsvWrKI3q4nmLYaiAwAAAJ7vnNgOAAAA; expires=Fri, 03-Sep-2021 20:29:21 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'fpc=AiQjbvDLFR5FiPCbJutjxyA; expires=Sun, 24-Oct-2021 16:46:24 GMT; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
-  'esctx=AQABAAAAAAD--DLA3VO7QrddgJg7WevrnQkyJSh4ZnzeCuQg8DiWR1kRtB0YweD1FSKyMLI-rE9LE-JFCbn5kcqzvi8gb7lwNas8Joozt7__EgBGnqGDomxBvStQWl6au6L3Zgg0QNFKtCjp0rQyJh1AfcsZ8LdOPvo5Fg2qbA03qBps9Vivd-x0yfLt-Xq5G-eQQXGnpuYgAA; domain=.login.microsoftonline.com; path=/; secure; HttpOnly; SameSite=None',
+  'esctx=AQABAAAAAAD--DLA3VO7QrddgJg7WevrAYM3xH6F4HVDyOibd78VdoAXf9FMkAbHqsz7lRUH8wgawVVZGsIin2mrQkwGQdVaKmi-Oyto2LvnYX3vZJgylU_81I2l3epu-HJlFUl8xAomNIFToKXkgfGlWosg8kFi9IiVym0hp9v2G6Uz1an36bgf_U5jy9TYK3G-ZNaRWjAgAA; domain=.login.microsoftonline.com; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
   'x-ms-gateway-slice=estsfd; path=/; secure; samesite=none; httponly',
   'Set-Cookie',
   'stsservicecookie=estsfd; path=/; secure; samesite=none; httponly',
   'Date',
-  'Wed, 04 Aug 2021 20:29:20 GMT',
+  'Fri, 24 Sep 2021 16:46:24 GMT',
   'Content-Length',
   '1753'
 ]);
 
 nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
-  .post('/azuretenantid/oauth2/v2.0/token', "client_id=azure_client_id&scope=https%3A%2F%2Fsanitized%2F&grant_type=client_credentials&x-client-SKU=msal.js.node&x-client-VER=1.2.0&x-client-OS=linux&x-client-CPU=x64&x-ms-lib-capability=retry-after, h429&x-client-current-telemetry=2|771,0|,&x-client-last-telemetry=2|0|||0,0&client-request-id=38dc57e9-90f2-4696-8dab-575d5e9b867f&client_secret=azure_client_secret")
+  .post('/azuretenantid/oauth2/v2.0/token', "client_id=azure_client_id&scope=https%3A%2F%2Fsanitized%2F&grant_type=client_credentials&x-client-SKU=msal.js.node&x-client-VER=1.3.1&x-client-OS=linux&x-client-CPU=x64&x-ms-lib-capability=retry-after, h429&x-client-current-telemetry=5|771,2,,,|,&x-client-last-telemetry=5|0|||0,0&client-request-id=b2cf1ae9-3d0e-4534-8d9a-cb24af5aa56f&client_secret=azure_client_secret&claims=%7B%22access_token%22%3A%7B%22xms_cc%22%3A%7B%22values%22%3A%5B%22CP1%22%5D%7D%7D%7D")
   .reply(200, {"token_type":"Bearer","expires_in":86399,"ext_expires_in":86399,"access_token":"access_token"}, [
   'Cache-Control',
   'no-store, no-cache',
@@ -93,19 +93,19 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
   'P3P',
   'CP="DSP CUR OTPi IND OTRi ONL FIN"',
   'x-ms-request-id',
-  '11b7b6b0-1a61-40b1-bbed-b518e1350700',
+  '84b61667-cf68-4ed9-837f-ea9e23233400',
   'x-ms-ests-server',
-  '2.1.11935.12 - WUS2 ProdSlices',
+  '2.1.12071.13 - WUS2 ProdSlices',
   'x-ms-clitelem',
   '1,0,0,,',
   'Set-Cookie',
-  'fpc=ArqRweA9seFErsvWrKI3q4nmLYaiBAAAAJ7vnNgOAAAA; expires=Fri, 03-Sep-2021 20:29:21 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'fpc=AgRTlwI41DFNqDkrVgDbl2XmLYaiAQAAAOD339gOAAAA; expires=Sun, 24-Oct-2021 16:46:24 GMT; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
   'x-ms-gateway-slice=estsfd; path=/; secure; samesite=none; httponly',
   'Set-Cookie',
   'stsservicecookie=estsfd; path=/; secure; samesite=none; httponly',
   'Date',
-  'Wed, 04 Aug 2021 20:29:20 GMT',
+  'Fri, 24 Sep 2021 16:46:24 GMT',
   'Content-Length',
   '1321'
 ]);
@@ -113,7 +113,7 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
 nock('https://endpoint', {"encodedQueryParams":true})
   .post('/$schemagroups/group-1/schemas/never-registered', {"type":"record","name":"User","namespace":"com.azure.schemaregistry.samples","fields":[{"name":"name","type":"string"},{"name":"favoriteNumber","type":"int"}]})
   .query(true)
-  .reply(404, {"Code":404,"Detail":"Schema group-1/never-registered does not exist. TrackingId:42203ade-ec98-4ad2-8639-bf0af80c98be_G6, SystemTracker:endpoint:$schemagroups/group-1/schemas/never-registered, Timestamp:2021-08-04T20:29:21"}, [
+  .reply(404, {"Code":404,"Detail":"Schema group-1/never-registered does not exist. TrackingId:66d3043d-e992-4c06-afeb-bb9a76406d90_G28, SystemTracker:endpoint:$schemagroups/group-1/schemas/never-registered, Timestamp:2021-09-24T16:46:25"}, [
   'Transfer-Encoding',
   'chunked',
   'Content-Type',
@@ -123,5 +123,5 @@ nock('https://endpoint', {"encodedQueryParams":true})
   'Strict-Transport-Security',
   'max-age=31536000',
   'Date',
-  'Wed, 04 Aug 2021 20:29:20 GMT'
+  'Fri, 24 Sep 2021 16:46:25 GMT'
 ]);