+++ sdk/core/core-client-paging-rest/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Other Changes
 
+- Updates package to work with the react native bundler. Browser APIs may still need to be pollyfilled for this package to run in react native. [PR #17783](https://github.com/Azure/azure-sdk-for-js/pull/17783)
+
 ## 1.0.0-beta.1 (2021-08-05)
 
 - First release of package, see README.md for details.